requirements.txt
pillow >= 6.1.0